requirements.txt
fastapi
uvicorn[standard]
pydantic
pytest
httpx
websockets
restrictedpython
google-generativeai
